requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2020.11.8
colorlover==0.3.0
cycler==0.10.0
decorator==4.4.1
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.17.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.0.1
plot-metric==0.0.6
progressbar==2.5
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
pyzmq==18.1.1
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
threadpoolctl==2.1.0
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
